+++ polaris-react/src/components/PageActions/tests/PageActions.test.tsx
@@ -55,7 +55,6 @@ describe('<PageActions />', () => {
       {
         content: 'Delete',
         destructive: true,
-        outline: true,
       },
     ];